Shift remaining code of TensorboardCallback()
