+++ experiments/sb3.py
@@ -104,48 +104,6 @@ if __name__ == "__main__":
     cb_chain = CallbackList([checkpoint_callback, eval_callback])
     model.learn(total_timesteps=train_timesteps, callback=cb_chain)
 
-    from stable_baselines3 import SAC
-    from stable_baselines3.common.callbacks import BaseCallback
-    from torch.utils.tensorboard import SummaryWriter
-
-
-        def _on_step(self) -> bool:
-
-            # Log scalar value (here a random variable)
-
-            return True
-
-        def __call__(self, a, b):
-            metrics = a["infos"][0]
-
-            # if len(metrics) != 0:
-            #    for item in list(metrics.keys())[2:]:
-            #        self.find_and_record(item)
-            # else:
-            #    self.iter += 1
-
-            # self.logger.dump(self.num_timesteps + self.iter*eval_timesteps)
-
-            # worker_number = self.env.label
-            run_number = 1
-
-            max_sumo_timestep = eval_timesteps  # self.env.sim_max_time
-            current_timestep = self.num_timesteps
-
-            total_sumo_timestep = current_timestep + run_number * max_sumo_timestep
-            # last_info = self.env.metrics[-1]
-
-            log_metric_names = list(metrics.keys())[1:]
-
-            for key, val in metrics.items():
-                # key += "/" + str(worker_number)
-                if type(val) is list or type(val) is np.ndarray:
-                    val = np.mean(val)
-                self.tb_writer.add_scalar(key, val, current_timestep)
-            # print("added")
-
-            print("tru")
-
     model.learn(
         total_timesteps=train_timesteps
     )  # , callback=TensorboardCallback(0))# callback=eval_callback)
